// schemas/FieldType.ts
export enum FieldType {
    Text = 'INPUT_TYPE_TEXT',
    Email = 'INPUT_TYPE_EMAIL',
    Phone = 'INPUT_TYPE_TEL',
    Number = 'INPUT_TYPE_NUMBER',
    Textarea = 'INPUT_TYPE_TEXTAREA',
    Checkbox = 'INPUT_TYPE_CHECKBOX',
    Radio = 'INPUT_TYPE_RADIO',
    Select = 'INPUT_TYPE_SELECT',
    FileUpload = 'INPUT_TYPE_FILE',
    Date = 'INPUT_TYPE_DATE',
    Time = 'INPUT_TYPE_TIME',
    DateTime = 'INPUT_TYPE_DATETIME_LOCAL',
}